moved download_single_submission to journalist_app/col.py
