+++ securedrop/tests/test_journalist.py
@@ -648,7 +648,7 @@ class TestJournalistApp(TestCase):
 
     def test_user_authorization_for_gets(self):
         urls = [url_for('main.index'), url_for('col.col', filesystem_id='1'),
-                url_for('download_single_submission',
+                url_for('col.download_single_submission',
                         filesystem_id='1', fn='1'),
                 url_for('account.edit')]
